requirements.txt
# Core ML
# NOTE: Install PyTorch CPU-only to save ~3GB disk space:
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# MIDI processing
pretty_midi>=0.2.10
miditoolkit>=1.0.1

# ONNX
onnx>=1.15.0
onnxruntime>=1.16.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
pandas>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
